pyproject.toml
[tool.poetry]
name = "youtube"
version = "0.1.0"
description = ""
authors = ["thanhphandev <thanh_dth225766@student.agu.edu.vn>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
python-telegram-bot = "^21.7"
python-dotenv = "^1.0.1"
openai = "^1.54.4"
youtube-transcript-api = "^0.6.3"
nltk = "^3.9.1"
groq = "^0.12.0"
pytube = "^15.0.0"
yt-dlp = "^2024.11.18"
google-api-python-client = "^2.153.0"
aiofiles = "^24.1.0"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
pydantic = "^2.10.2"
fastapi = "^0.115.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
